+++ source/driver/api/in/src/main/java/ee/qrental/driver/api/in/request/DriverUpdateRequest.java
@@ -1,6 +1,7 @@
 package ee.qrental.driver.api.in.request;
 
 import ee.qrental.common.core.in.request.AbstractUpdateRequest;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -25,6 +26,6 @@ public class DriverUpdateRequest extends AbstractUpdateRequest {
   private LocalDate driverLicenseExp;
   private String taxiLicense;
   private String address;
-  private Long deposit;
+  private BigDecimal deposit;
   private String comment;
 }
